Move external connection factory back
